fix: type changed

+++ src/main/java/ai/synthai/businessbackend/application/dto/TranscriptionResponseDto.java
@@ -10,6 +10,6 @@ public record TranscriptionResponseDto<T>(
         Status status,
         T transcriptionAnalysis,
         Category category,
-        Float duration,
+        int duration,
         Language language) {
 }
